# Optimize TestSuite comparison macro compile times.