--- a/src/Corrade/TestSuite/CMakeLists.txt
+++ b/src/Corrade/TestSuite/CMakeLists.txt
@@ -32,6 +32,7 @@ set(CorradeTestSuite_SRCS
     Comparator.cpp
     Tester.cpp
 
+    Compare/Container.cpp
     Compare/File.cpp
     Compare/FileToString.cpp
     Compare/FloatingPoint.cpp