I regularly get angry at how long large tests take to compile. Well, I guess relatively, building a heavy 6000-line `GltfImporterTest.cpp` in ~4.5 seconds is an unachievable feat in Some Other Codebases. But still.

Profiling using Clang `-ftime-trace`, the biggest offenders are:

- parsing the `<sstream>` include, about 160 ms / 3.5%  -- depends on making Debug stream-free, which depends on ... making my own float printers
- parsing `<cmath>`, about 40 ms / 1% -- can't really do much about this, and the more I optimize, the more this will stand out
- template instantiations, about 1000 ms / 20%, of which the biggest chunks are:
  - [ ] `TestSuite::Compare::Container<T>::printMessage()`, about 500 ms / 10% -- put it into a non-templated base. These are not so many but quite large.
    - [ ] The comparator also does the comparisons twice on error, which is fine at runtime but adds even more pain to compile times. Why not just remember the first different item instead, instead of looping through everything again?!
  - [x] `TestSuite::Comparator<T>::printMessage()` -- put it into a non-templated base. These are tiny but quite many, so the template base might not help as much.
  - [ ] Others, visible once the two above are gone?